All Working Except Cooking and Equipment agents

+++ src/main/java/ru/hse/jade/sample/model/products_on_stock_list/ProductOnStock.java
@@ -9,7 +9,7 @@ public class ProductOnStock {
     public String prod_item_company;
     public String prod_item_unit;
     public double prod_item_quantity;
-    public int prod_item_cost;
+    public double prod_item_cost;
     public Date prod_item_delivered;
     public Date prod_item_valid_until;
 }
